backend/material_extraction/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
PyPDF2==3.0.1
pdfplumber==0.11.4
python-docx==1.1.2
python-pptx==1.0.2
Pillow==10.4.0
pytesseract==0.3.13
python-magic==0.4.27
aiofiles==24.1.0
pydantic==2.9.0
spacy==3.7.6
